+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add saving of things as shortcuts (#17)
 * Save selection with page titles for filenames
 * Add filename conflict action option for Chrome
+* Add reset to default button in options
 
 # 1.5.2